pyproject.toml
[tool.poetry]
name = "seo-dp"
version = "0.1.0"
description = "SEO Data Platform for extracting data from Google Search Console and Google Analytics 4"
authors = ["JR Oakes <jroakes@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
google-auth = "^2.35.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.149.0"
google-analytics-data = "^0.18.12"
jinja2 = "^3.1.4"
beautifulsoup4 = "^4.12.3"
trafilatura = "^1.12.2"
scrapingbee = "^2.0.1"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
apscheduler = "^3.10.4"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

